+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -10,7 +10,7 @@ import lombok.Value;
 public class Genre extends BaseEntity {
     private String name;
 
-    public Genre(int id, String name) {
+    public Genre(long id, String name) {
         super(id);
         this.name = name;
     }
